getting rid of Hybrid Encryption
