+++ crate/utils/src/crypto/mod.rs
@@ -2,7 +2,6 @@ pub mod cover_crypt;
 pub mod dh_shared_keys;
 pub mod elliptic_curves;
 pub mod generic;
-pub mod hybrid_encryption;
 pub mod password_derivation;
 pub mod rsa;
 pub mod secret;
